+++ apps/journeys-admin/src/components/Editor/Canvas/AddBlockToolbar/Button/Button.spec.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render } from '@testing-library/react'
 
-import { ThemeProvider } from '../../../ThemeProvider'
+import { ThemeProvider } from '../../../../ThemeProvider'
 
 import { Button } from '.'
