+++ apps/journeys-admin/src/components/Editor/ControlPanel/ControlPanel.tsx
@@ -36,14 +36,6 @@ const Attributes = dynamic(
     ).then((mod) => mod.Attributes),
   { ssr: false }
 )
-const AddBlockToolbar = dynamic(
-  async () =>
-    await import(
-      /* webpackChunkName: "Editor/ControlPanel/AddBlockToolbar" */
-      '../Canvas/AddBlockToolbar'
-    ).then((mod) => mod.AddBlockToolbar),
-  { ssr: false }
-)
 const CardTemplateDrawer = dynamic(
   async () =>
     await import(
@@ -296,14 +288,6 @@ export function ControlPanel(): ReactElement {
             <Attributes selected={selected} step={selectedStep} />
           )}
         </TabPanel>
-        {/* <TabPanel
-          name="control-panel"
-          value={activeTab}
-          index={2}
-          unmountUntilVisible
-        >
-          <AddBlockToolbar />
-        </TabPanel> */}
       </Stack>
     </Stack>
   )
